+++ asynctor/compat.py
@@ -1,14 +1,41 @@
+"""
+This module is for those that support both (python >=3.9,<3.11) and (python >= 3.11)
+"""
+
 from __future__ import annotations
 
 import os
 import sys
+import warnings
+from typing import Any
+
+__all__ = ("chdir", "StrEnum", "load_toml", "Self")
 
 if sys.version_info >= (3, 11):
     from contextlib import chdir
     from enum import StrEnum
+    from typing import Self
+
+    from tomllib import loads as load_toml
 else:
     import contextlib
     from enum import Enum
+
+    from typing_extensions import Self
+
+    def load_toml(content: str) -> dict[str, Any]:
+        try:
+            import tomli
+        except ImportError:
+            tip = (
+                "tomli is required for Python<3.11, you can install it by:"
+                "\n\n  pip install tomli"
+                '\n\nOr:\n\n  pip install "asynctor[toml]"\n'
+            )
+            warnings.warn(tip, stacklevel=2)
+            raise
+
+        return tomli.loads(content)
 
     class chdir(contextlib.AbstractContextManager):  # Copied from source code of Python3.13
         """Non thread-safe context manager to change the current working directory."""